Extracted generic render pass to render cookie atlas to a new module
